+++ derive-input/src/internal/db.rs
@@ -1,0 +1,3 @@
+pub mod table;
+
+pub mod column;
