+++ derive-input/src/internal/dsl/wrapper.rs
@@ -2,12 +2,13 @@ use super::{path, wrap, wrapped};
 use crate::api::dsl::wrapper::{Wrap, Wrapped, WrapperType};
 use crate::api::rust::{RustField, RustStruct};
 use ident_case::RenameRule;
+use proc_macro2::TokenStream;
 use quote::quote;
 use quote::{ToTokens, format_ident};
 use spacetime_bindings_macro_input::{
     match_meta, sats::SatsField, sym::name, util::check_duplicate,
 };
-use syn::{parse_str, Error, Ident, Path, Type};
+use syn::{Error, Ident, Path, Type, parse_str};
 
 impl WrapperType {
     pub(in crate::internal) fn try_parse(
@@ -109,13 +110,20 @@ fn get_wrapper_impl(
 ) -> Box<str> {
     let struct_name = format_ident!("{struct_name}");
     let wrapper_struct_name = format_ident!("{wrapper_struct_name}");
-    let wrapped_type: Type = parse_str(wrapped_type_name_or_path).expect(&format!("Expected to parse {wrapped_type_name_or_path} as Type in get_wrapper_impl!"));
+    let wrapped_type: Type = parse_str(wrapped_type_name_or_path).expect(&format!(
+        "Expected to parse {wrapped_type_name_or_path} as Type in get_wrapper_impl!"
+    ));
     let field_name = format_ident!("{field_name}");
 
     let default_impl;
 
     if wrapped_type_name_or_path.starts_with("Option <") {
-        let wrapped_type_name_or_path: Type= parse_str(&wrapped_type_name_or_path.replace("Option <", "").replace(">", "")).unwrap();
+        let wrapped_type_name_or_path: Type = parse_str(
+            &wrapped_type_name_or_path
+                .replace("Option <", "")
+                .replace(">", ""),
+        )
+        .unwrap();
 
         default_impl = quote! {
             Some(#wrapped_type_name_or_path::default())
@@ -170,5 +178,44 @@ fn get_wrapper_impl(
                 self.value.clone()
             }
         }
-    }.to_string().into()
+    }
+    .to_string()
+    .into()
+}
+
+impl WrapperType {
+    pub(in crate::internal) fn map_to_wrapped_type(value: &Wrap) -> Type {
+        parse_str(&value.wrapped_type_name_or_path).expect(&format!(
+            "Failed to parse {} as Ident in WrapperType::map_to_wrapped_type.",
+            &value.wrapped_type_name_or_path
+        ))
+    }
+
+    pub(in crate::internal) fn map(value: &WrapperType) -> Type {
+        match value {
+            WrapperType::Wrap(w) => parse_str(&w.wrapper_struct_name).expect(&format!(
+                "Failed to parse {} as Ident in WrapperType::map_to_wrapper_type for WrapperType::Wrap.",
+                &w.wrapper_struct_name
+            )),
+            WrapperType::Wrapped(w) => parse_str(&w.wrapper_struct_name_or_path).expect(&format!(
+                "Failed to parse {} as Path in WrapperType::map_to_wrapper_type for WrapperType::Wrapped.",
+                &w.wrapper_struct_name_or_path
+            )),
+        }
+    }
+}
+
+pub(in crate::internal) fn wrapper_type_into_option(
+    column_name: &Ident,
+    wrapper_type_name_or_path: &Type,
+) -> TokenStream {
+    let column_option_name = &format_ident!("{column_name}_option");
+    quote! {
+        let #column_name = #column_name.into();
+        let mut #column_option_name = None;
+        if #column_name.is_some() {
+            #column_option_name = Some(Into::<#wrapper_type_name_or_path>::into(#column_name.unwrap()).value());
+        }
+        let #column_name = #column_option_name;
+    }
 }
