+++ derive-input/src/internal/rust/table.rs
